scripts/knowledge-base/requirements.txt
# Knowledge Base Dependencies

# OpenAI API for embeddings
openai>=1.12.0

# Qdrant vector database client
qdrant-client>=1.7.0

# Environment variable management
python-dotenv>=1.0.0

# Core utilities
requests>=2.31.0
typing-extensions>=4.9.0
tenacity>=8.2.0

# For text processing
tiktoken>=0.5.2  # For token counting 